invest type form update
