+++ invest_planner/base/forms.py
@@ -4,6 +4,7 @@ Base App Forms
 from django import forms
 from .models import InvestmentTag, IncomeTag, ExpenseTag
 from .models import Investment
+from .utils import get_central_bank_rate
 
 
 class InvestmentForm(forms.ModelForm):
@@ -15,20 +16,22 @@ class InvestmentForm(forms.ModelForm):
         )
     )
 
+    rate_type = forms.ChoiceField(choices=[('', '------')], required=False)
+    rate_percentage = forms.DecimalField(max_digits=5, decimal_places=2, required=False, initial=0)
+
     class Meta:
         model = Investment
-        fields = ['title', 'starting_amount', 'number_of_years', 'investment_type',
+        fields = ['title', 'starting_amount', 'number_of_years',
                   'return_rate', 'rate_type', 'rate_percentage',
                   'additional_contribution', 'active', 'starting_date', 'tags']
 
-    def clean(self):
-        cleaned_data = super().clean()
-        investment_type = cleaned_data.get('investment_type')
-        rate_percentage = cleaned_data.get('rate_percentage')
-        if investment_type == 'pos' and rate_percentage is None:
-            self.add_error(
-                'rate_percentage', 'Por favor, forneça a porcentagem da taxa para investimentos pós-fixados.')
-        return cleaned_data
+    def __init__(self, *args, **kwargs):
+        super(InvestmentForm, self).__init__(*args, **kwargs)
+        central_bank_rates = get_central_bank_rate()
+        rate_choices = [(tax['name'], f"{tax['name'].upper()} ({tax['rate']}%)") for tax in central_bank_rates['taxes']]
+        self.fields['rate_type'].choices += rate_choices
+        # Initialize rate_percentage to 0 and set it as inaccessible if rate_type is empty
+        self.fields['rate_percentage'].widget.attrs['readonly'] = True
 
 
 class InvestmentTagForm(forms.ModelForm):
