+++ invest_planner/base/templates/base/navbar.html
@@ -29,13 +29,13 @@
                 </li>
                 <li class="nav-item dropdown">
                     <a class="nav-link dropdown-toggle" href="#" id="navbarDropdown" role="button"
-                        data-bs-toggle="dropdown" aria-haspopup="true" aria-expanded="false">Placeholder</a>
+                        data-bs-toggle="dropdown" aria-haspopup="true" aria-expanded="false">Menu</a>
                     <div class="dropdown-menu" aria-labelledby="navbarDropdown">
-                        <a class="dropdown-item" href="{% url "account_email" %}">account_email</a>
+                        <!-- <a class="dropdown-item" href="{ url "account_email" %}">account_email</a> -->
                         <a class="dropdown-item" href="{% url "tag-list" %}">Manage Tags</a>
                         <a class="dropdown-item" href="{% url "investment-rates" %}">investment-rates</a>
                         <div class="dropdown-divider"></div>
-                        <a class="dropdown-item" href="#">Separated link</a>
+                        <a class="dropdown-item" href="#">placeholder</a>
                     </div>
                 </li>
                 {% endif %}
