+++ invest_planner/base/templates/tags/tag_list.html
@@ -37,7 +37,7 @@
             </li>
         {% endfor %}
     </ul>
-    <h2>Tags (Temporary)</h2>
+    <h2>All Tags (Temporary)</h2>
     <ul>
         {% for tag in tags %}
             <li>
